GSSG-183 MyPageProfile 삭제
- 재활용 될 건덕지? 가 없음.
